startstop/requirements.txt
kubernetes
pyyaml
